requirements.txt
pytest-playwright
bs4
psycopg2-binary
